Add support for (at least) the PiTFT touchscreen.
We build most of the touchscreen support directly into the kernel,
because it's more convenient.  But the fbtft device has to be loaded
late in kernel init, and most of the time we're not going to use it,
so special-case that as a module.

The b2g.sh and init.rc changes implement a hack to get b2g to use the
right framebuffer.  The gonk level hard-codes /dev/graphics/fb0
everywhere, so we check the persist.fbtft property at b2g start time
to determine what the user wants.  Then we make a link from
/dev/graphics/fb0 to the right framebuffer device, if necessary.